bckend/requirements.txt
Flask
flask-cors
python-dotenv
langchain
langchain-google-genai
markdown2